Enable plugin URLs for Dashboard List items
